refactor(apiLayers): merge into single file for consistency
